# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: StockPilot - Inventory & POS SaaS
  description: '

    ## StockPilot API - Inventory & POS SaaS


    Multi-tenant inventory management, point-of-sale, finance tracking, and AI insights

    for small businesses.


    ### Authentication

    All API endpoints require a Bearer token in the Authorization header:

    `Authorization: Bearer <access_token>`


    Include the business ID in the X-Business-Id header:

    `X-Business-Id: <business_id>`


    ### Rate Limiting

    - Default: 100 requests per minute per user

    - Auth endpoints: 5 requests per minute per IP


    ### Support

    For issues and feature requests, contact support.

    '
  version: 1.0.0
paths:
...
  /api/v1/returns:
    post:
      tags:
      - returns
      summary: Create Return
      description: 'FR-14: full/partial return -> stock up + financial adjustment + audit.'
      operationId: create_return_api_v1_returns_post
      security:
      - HTTPBearer: []
      parameters:
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - returns
      summary: List Returns
      operationId: list_returns_api_v1_returns_get
      security:
      - HTTPBearer: []
      parameters:
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReturnOut'
                title: Response List Returns Api V1 Returns Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ReturnCreate:
      properties:
        sale_id:
          type: integer
          title: Sale Id
        reason:
          type: string
          minLength: 1
          title: Reason
        items:
          items:
            $ref: '#/components/schemas/ReturnItemIn'
          type: array
          minItems: 1
          title: Items
      type: object
      required:
      - sale_id
      - reason
      - items
      title: ReturnCreate
    ReturnItemIn:
      properties:
        sale_item_id:
          type: integer
          title: Sale Item Id
        quantity:
          type: integer
          exclusiveMinimum: 0.0
          title: Quantity
      type: object
      required:
      - sale_item_id
      - quantity
      title: ReturnItemIn
    ReturnOut:
      properties:
        id:
          type: integer
          title: Id
        sale_id:
          type: integer
          title: Sale Id
        reason:
          type: string
          title: Reason
        refund_amount:
          type: number
          title: Refund Amount
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - sale_id
      - reason
      - refund_amount
      - created_at
      title: ReturnOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer